We kunnen een correlatie in stappen uitvoeren! In WIDTH+HEIGHT^2 stappen om precies te zijn. We hebben nog geen hamming filter

+++ frame.cpp
@@ -3,8 +3,8 @@
 
 void frame_fill(px_t* buf, u16 x, u16 y, px_t px) {
 
-	const u16 newx = (x * SMALL_WIDTH) / WIDTH;
-	const u16 newy = (y * SMALL_HEIGHT) / HEIGHT;
+	const u32 newx = (x * SMALL_WIDTH) / WIDTH;
+	const u32 newy = (y * SMALL_HEIGHT) / HEIGHT;
 
 	if( x % (WIDTH / SMALL_WIDTH) == 0 && y % (HEIGHT / SMALL_HEIGHT) == 0){
 		buf[newx + newy * SMALL_WIDTH] = px;
